Autocannon Boogaloo
Added placeholder autocannon. Has very basic sprites and one of the issues is dissapearing sprite when firing. This is due to the ammo counter breaking due to it having a box of 400 rounds.

The map was expanded northward and now has light at all times and a shooting range full of actually terrifying enemies.